Refactor: Move feature-specific root modules
